fixed ES params in settings.cfg
removed import_ES_AWS.py and added Opensearch support in import_AWS.py
added opensearchpy in setup.py
fixed some adhoc scripts
