+++ scripts/geonames/import_ES.py
@@ -12,7 +12,15 @@ from future import standard_library
 standard_library.install_aliases()
 
 import traceback
+import argparse
+from datetime import datetime
+import urllib3
+
+import boto3
 from elasticsearch import Elasticsearch
+from opensearchpy import OpenSearch, AWSV4SignerAuth
+
+urllib3.disable_warnings()
 
 
 '''
@@ -28,8 +36,8 @@ country code      : ISO-3166 2-letter country code, 2 characters
 cc2               : alternate country codes, comma separated, ISO-3166 2-letter country code, 60 characters
 admin1 code       : fipscode (subject to change to iso code), see exceptions below, see file admin1Codes.txt for
                     display names of this code; varchar(20)
-admin2 code       : code for the second administrative division, a county in the US
-                    see file admin2Codes.txt;varchar(80)
+admin2 code       : code for the second administrative division, a county in the US,
+                    see file admin2Codes.txt; varchar(80)
 admin3 code       : code for third level administrative division, varchar(20)
 admin4 code       : code for fourth level administrative division, varchar(20)
 population        : bigint (8 byte int)
@@ -127,9 +135,6 @@ MAPPING = {
 }
 
 
-ES_URL = 'http://localhost:9200'
-
-
 def get_admin_dict(admin_file):
     d = {}
     with open(admin_file) as f:
@@ -150,27 +155,46 @@ def get_country_dict(country_file):
     return d
 
 
-def parse(csv_file):
-
-    # get adm dicts
-    adm1 = get_admin_dict('admin1CodesASCII.txt')
-    adm2 = get_admin_dict('admin2Codes.txt')
-
-    # get country dict
-    cntries = get_country_dict('countryInfo.txt')
-
-    # get ElasticSearch connection
-    es = Elasticsearch(hosts=[ES_URL])
-    es.indices.create(INDEX, {'mappings': MAPPING}, ignore=400)
+def create_geonames_mapping(es):
+    """
+    :param es: python elasticsearch's es object
+    :return:
+    """
+    mapping = {
+        "mappings": MAPPING
+    }
+    es.indices.create(INDEX, mapping, ignore=400)
     print('%s index created!!' % INDEX)
 
-    # iterate and index
+
+def parse(es, csv_file, start_position=0):
+    """
+    index geonames docs into the geonames index
+    :param csv_file: string
+    :param es: elasticsearch-py's es object
+    :param start_position: file position to start loop at (if we don't want to start from the beginning)
+    :return:
+    """
+    adm1 = get_admin_dict('admin1CodesASCII.txt')  # get adm dicts
+    adm2 = get_admin_dict('admin2Codes.txt')
+    cntries = get_country_dict('countryInfo.txt')  # get country dict
+
     line_number = 0
+    line_position = 0  # tracking line_position in case we don't want to start from the beginning
     try:
         with open(csv_file) as f:
-            for line in f:
-                row = dict(list(zip(FIELD_NAMES, line.strip().split('\t'))))
+            if start_position > 0:
+                f.seek(start_position)  # if we do not want to start from beginning
+
+            while True:
+                line_position = f.tell()  # change current line position
+                line = f.readline()
+                if not line:
+                    break
+
+                row = dict(list(zip(FIELD_NAMES, line.strip().split('\t'))))  # creates a dictionary per row
                 line_number += 1
+
                 for k in row:
                     if row[k] == '':
                         if k in ('alternatename', 'cc2'):
@@ -212,16 +236,76 @@ def parse(csv_file):
                                 adm1_code, row['admin2_code'])
                             row['admin2_name'] = adm2.get(adm2_code, [None])[0]
 
-                # index
                 es.index(index=INDEX, id=row['geonameid'], body=row)
                 if line_number % 10000 == 0:
-                    print('%d documents ingested into %s' % (line_number, INDEX))
+                    print('%d documents ingested at %s, position: %d' % (line_number, datetime.now().isoformat(),
+                                                                         line_position))
     except Exception as e:
         traceback.print_exc()
-        print(("line_number: %d" % line_number))
-        raise
+        print('line_number: %d' % line_number)
+        print('failed position %d' % line_position)
+        print('failed at %s' % datetime.now().isoformat())
+        raise e
 
 
-if __name__ == "__main__":
-    csv_file = 'allCountries.txt'
-    parse(csv_file)
+if __name__ == '__main__':
+    parser = argparse.ArgumentParser(description='Create and populate geonames index')
+
+    parser.add_argument('--engine', type=str, default="elasticsearch", help="elasticsearch/opensearch")
+    parser.add_argument('--es-url', action='store', default='http://localhost:9200', help="ES endpoint")
+    parser.add_argument('--aws', action='store_true', default=False, help="AWS managed Elasticsearch/OpenSearch")
+    parser.add_argument('--region', action='store', default='us-west-1', help="AWS region")
+    parser.add_argument('--verify-certs', action='store_true', default=False, help='verify SSL if using https')
+    parser.add_argument('--start-position', type=int, default=0, help='file position to start at')
+
+    args = parser.parse_args()
+
+    region = args.region  # us-west-1
+    engine = args.engine
+    aws = args.aws
+    es_url = args.es_url
+
+    if engine == "opensearch":
+        if aws is True or "es.amazonaws.com" in es_url:
+            from elasticsearch.connection import RequestsHttpConnection
+
+            credentials = boto3.Session().get_credentials()
+            auth = AWSV4SignerAuth(credentials, region)
+            grq_es = OpenSearch(
+                [es_url],
+                http_auth=auth,
+                connection_class=RequestsHttpConnection,
+                use_ssl=True,
+                verify_certs=args.verify_certs,
+                ssl_show_warn=False,
+                timeout=30,
+                max_retries=10,
+                retry_on_timeout=True,
+            )
+        else:
+            grq_es = OpenSearch([es_url])
+    else:
+        if aws is True or "es.amazonaws.com" in es_url:
+            from opensearchpy.connection import RequestsHttpConnection
+
+            credentials = boto3.Session().get_credentials()
+            auth = AWSV4SignerAuth(credentials, region)
+            grq_es = Elasticsearch(
+                [es_url],
+                http_auth=auth,
+                connection_class=RequestsHttpConnection,
+                use_ssl=True,
+                verify_certs=args.verify_certs,
+                ssl_show_warn=False,
+                timeout=30,
+                max_retries=10,
+                retry_on_timeout=True,
+            )
+        else:
+            grq_es = Elasticsearch([es_url])
+
+    print("script start time: %s" % datetime.now().isoformat())
+    create_geonames_mapping(grq_es)  # create geonames index with mapping
+
+    countries_csv_file = 'allCountries.txt'
+    parse(grq_es, countries_csv_file, args.start_position)
